Support static linking for Ruby
This approach essentially follows the logic outlined by @alexcrichton in
[this
issue](https://github.com/steveklabnik/ruby-sys/pull/16#issuecomment-271412827).

I have tested it on Linux against `ruru` and a few other gems, and things seem
to be working. The only hiccup I found was that we needed to specify the `LIBS`
which were used from Ruby. With the approach on #16, this was not neccesary. I
am not sure why. Does anyone have any insight to this?

Any feedback would be awesome, especially if you are a Mac user.

1. Clone ruru: `https://github.com/d-unseductable/ruru`
1. Update Cargo.toml in ruru: `ruby-sys = { git = "https://github.com/ianks/ruby-sys" }`
1. Run the tests: `RUBY_STATIC=1 cargo test`